requirements.txt
FlowCal
numpy
seaborn
matplotlib
pandas
biopython
tqdm